requirements.txt
playwright
pytest
pytest-bdd
pytest-xdist
python-dotenv
faker
pytest-html
openpyxl